Updating the Project

+++ Models/Domain/InvoiceDetail.cs
@@ -5,7 +5,7 @@ namespace EmployeeClient.Models.Domain
     public class InvoiceDetail
     {
         public int InvoiceDetailId { get; set; }
-        public decimal Qty { get; set; }
+        public int Qty { get; set; }
         public decimal UnitPrice { get; set; }
         public decimal SubTotal { get; set; }
         [ForeignKey("Product")]
